requirements.txt
django
six
djangorestframework
dateutils
